refactor: Article, Comment command 패키지 경로 변경
- command 와 result 로 패키지 분리
