+++ src/main/java/com/mobble/mobbleserver/application/article/command/UpdateArticleCommand.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.application.article.command.request;
+package com.mobble.mobbleserver.application.article.command;
 
 public record UpdateArticleCommand(
         Long memberId,
