+++ src/main/java/com/mobble/mobbleserver/application/article/port/provided/ArticleCreatePort.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.application.article.port.provided;
 
-import com.mobble.mobbleserver.application.article.command.request.CreateArticleCommand;
+import com.mobble.mobbleserver.application.article.command.CreateArticleCommand;
 import com.mobble.mobbleserver.domain.article.Article;
 
 public interface ArticleCreatePort {
