+++ src/main/java/com/mobble/mobbleserver/application/article/port/provided/ArticleUpdatePort.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.application.article.port.provided;
 
-import com.mobble.mobbleserver.application.article.command.request.UpdateArticleCommand;
+import com.mobble.mobbleserver.application.article.command.UpdateArticleCommand;
 import com.mobble.mobbleserver.domain.article.Article;
 
 public interface ArticleUpdatePort {
