+++ src/main/java/com/mobble/mobbleserver/application/comment/command/UpdateCommentCommand.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.application.comment.command.request;
+package com.mobble.mobbleserver.application.comment.command;
 
 public record UpdateCommentCommand(
         Long memberId,
