+++ src/main/java/com/mobble/mobbleserver/application/comment/port/provided/CommentCreatePort.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.application.comment.port.provided;
 
-import com.mobble.mobbleserver.application.comment.command.request.CreateReplyCommentCommand;
-import com.mobble.mobbleserver.application.comment.command.request.CreateRootCommentCommand;
+import com.mobble.mobbleserver.application.comment.command.CreateReplyCommentCommand;
+import com.mobble.mobbleserver.application.comment.command.CreateRootCommentCommand;
 import com.mobble.mobbleserver.domain.comment.Comment;
 
 public interface CommentCreatePort {
