+++ src/main/java/com/mobble/mobbleserver/application/comment/port/provided/CommentUpdatePort.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.application.comment.port.provided;
 
-import com.mobble.mobbleserver.application.comment.command.request.UpdateCommentCommand;
+import com.mobble.mobbleserver.application.comment.command.UpdateCommentCommand;
 import com.mobble.mobbleserver.domain.comment.Comment;
 
 public interface CommentUpdatePort {
