+++ src/main/java/com/mobble/mobbleserver/application/comment/service/CommentModifyService.java
@@ -3,9 +3,9 @@ package com.mobble.mobbleserver.application.comment.service;
 import com.mobble.mobbleserver.application.article.error.ArticleBusinessError;
 import com.mobble.mobbleserver.application.article.port.required.ArticleReadPort;
 import com.mobble.mobbleserver.application.clubMember.port.required.ClubMemberReadPort;
-import com.mobble.mobbleserver.application.comment.command.request.CreateReplyCommentCommand;
-import com.mobble.mobbleserver.application.comment.command.request.CreateRootCommentCommand;
-import com.mobble.mobbleserver.application.comment.command.request.UpdateCommentCommand;
+import com.mobble.mobbleserver.application.comment.command.CreateReplyCommentCommand;
+import com.mobble.mobbleserver.application.comment.command.CreateRootCommentCommand;
+import com.mobble.mobbleserver.application.comment.command.UpdateCommentCommand;
 import com.mobble.mobbleserver.application.comment.error.CommentBusinessError;
 import com.mobble.mobbleserver.application.comment.port.provided.CommentCreatePort;
 import com.mobble.mobbleserver.application.comment.port.provided.CommentDeletePort;
